拆分 AI 编排辅助流程
